+++ app/project/bootstrappers/http/sessions/Session.php
@@ -48,7 +48,7 @@ class Session extends BaseBootstrapper
     {
         $this->loadConfig();
 
-        switch($this->environment->getVariable("SESSION_HANDLER"))
+        switch($this->environment->getVar("SESSION_HANDLER"))
         {
             case CacheSessionHandler::class:
                 $handler = new CacheSessionHandler($this->getCacheBridge($container), $this->config["lifetime"]);
@@ -74,7 +74,7 @@ class Session extends BaseBootstrapper
      */
     private function getCacheBridge(IContainer $container)
     {
-        switch($this->environment->getVariable("SESSION_CACHE_BRIDGE"))
+        switch($this->environment->getVar("SESSION_CACHE_BRIDGE"))
         {
             case ArrayBridge::class:
                 return new ArrayBridge();